- fixed warnings - streamlined naming relations between source- and resource files.
--HG--
extra : convert_revision : svn%3Acdae58ac-7104-0410-addd-a5e18705bff5/trunk%4095

+++ Extensions/JRandom.cpp
@@ -85,7 +85,7 @@ short jrLinearShort(register jrPtr theJRPtr,
 		j = (NRANDOMSLOTS*theJRPtr->value) / NDISTINCTRANDOMVALUES;
 		theJRPtr->value = theJRPtr->seeds[j];
 		theJRPtr->seeds[j] = theJRPtr->dummy = NEXTVAL(theJRPtr->dummy);
-		if (theJRPtr->value < partitionSize * interval) {
+		if (theJRPtr->value < (unsigned long)(partitionSize * interval)) {
 			return lowest + theJRPtr->value/partitionSize;
 		}
 	}
